requirements.txt

openai>=1.40.0
pyyaml>=6.0.1
python-slugify>=8.0.4
google-analytics-data>=0.18.12
google-auth>=2.32.0
google-auth-oauthlib>=1.2.1
requests>=2.32.3
